+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Dev
 
+### Enhancements
+
+- Optimize `Aja.Vector.flat_map/2` and `Aja.Enum.flat_map/2` when used for
+  filtering
+- Bump `ex_doc` to the latest version
+
 ## v0.6.5 (2024-04-26)
 
 ### Enhancements